Bioconductor commit 3
Shortened hammingCompare, combineExpression, and clonaOverlap to <= 50 lines. As of now, combineTCR, combineBCR and lengthContig are the only functions > 50 lines

Added Utils.R for non-exported functions. Removed description for hammingCompare to make it internal function, but left it on the seuratFunction.R

Still need to add examples for the seurat/SCE-based functions - remove the dontrun{} commands