Bioconductor 4
Reduced size of combineTCR, combineBCR and lengthContig() to < 50 lines. Modularized chunks of code for these functions in the utils.R.

Annotated the utils.R functions.

Still need to develop examples for:
 alluvialClonotypes.Rd
        combineBCR
        combineExpression
        expression2List
        highlightClonotypes
        StartracDiversity
(This will removed last error of 80% of export, remove ntoes for donttest{})

Ready to submit to bioconductor after that list.